Explicit cast to avoid narrowing warning.

+++ opm/simulators/wells/MultisegmentWellSegments.cpp
@@ -815,7 +815,7 @@ pressureDropValve(const int seg,
     const EvalWell friction_pressure_loss =
         mswellhelpers::frictionPressureLoss(additional_length, diameter, area, roughness, density, mass_rate, visc);
 
-    const ValveUDAEval uda_eval {summary_state, this->well_.name(), segment_set[seg].segmentNumber()};
+    const ValveUDAEval uda_eval {summary_state, this->well_.name(), static_cast<std::size_t>(segment_set[seg].segmentNumber())};
     const double area_con = valve.conCrossArea(uda_eval);
     const double cv = valve.conFlowCoefficient();
